ESN-Database/requirements.txt
aniso8601==6.0.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
mysql-connector==2.1.6
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.1
